+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `asynqmon workers` was added to list all running workers information
+
 ## [0.4.0] - 2020-02-13
 
 ### Changed